refactor(client): split livekit_voice_provider AV controls into part (2/7)
Second god-module split per .claude/plans/generic-growing-bentley.md.
Pure reorganization — zero behavior change, public API identical.

Layout:
- providers/livekit_voice/livekit_voice_provider.dart — facade: state,
  joinChannel/leaveChannel/dispose, room listeners, peer sync, audio
  level polling, token fetch.
- providers/livekit_voice/livekit_voice_av_controls.dart — mixin: mic,
  camera, screen share, video bitrate/fps, auto-quality + the friendly
  media-error mapper.
- providers/livekit_voice_provider.dart — re-export shim for existing imports.

Same mixin-on-StateNotifier pattern as PR #771. Mixin reaches the
shared _room / _disposed / _wasMutedBeforeDeafen fields back through
abstract accessors so the field declarations stay on the facade.

All 1305 client tests pass at the new layout.
